[#22] Added Previous Location to avoid Phase through on Collision
This means that provided the movement does not mean a collision no longer occurred or that they do not collide with the first thing in their path (or linked to the first thing, like a brick wall would be fine because they would be pushed back) then the rigidbody is moving the player back to the correct position.

The fallback right now is previous position if it fails which could lead to weird behaviour.